Excepciones

+++ src/main/java/com/proyecto/registro/services/UsuarioService.java
@@ -2,6 +2,7 @@ package com.proyecto.registro.services;
 
 import com.proyecto.registro.dao.UsuarioDao;
 import com.proyecto.registro.dto.UsuarioInDTO;
+import com.proyecto.registro.exceptions.UsuarioNotFoundException;
 import com.proyecto.registro.mappers.UsuarioInDTOToUsuario;
 import com.proyecto.registro.persistences.models.UserStatus;
 import com.proyecto.registro.persistences.models.Usuario;
@@ -26,19 +27,30 @@ public class UsuarioService {       //Toda la lógica de negocio, como crear nue
         return this.usuarioDao.selectAllUsuarios();
     }
     public Usuario getUsuario(Integer id) {
-        return this.usuarioDao.selectUsuarioById(id);
+        return this.usuarioDao.selectUsuarioById(id)
+                .orElseThrow(() -> new UsuarioNotFoundException(
+                        "Usuario con id [%s] no encontrado".formatted(id)
+                ));
     }
     public Usuario addUsuario(UsuarioInDTO usuarioInDTO) {
         Usuario usuario = mapper.map(usuarioInDTO);           // Esto convierte UsuarioInDTO en Usuario
         return this.usuarioDao.insertUsuario(usuario);
     }
     public void updateUsuario(Integer id, Usuario updateRequest) {
-        Usuario usuario = this.usuarioDao.selectUsuarioById(id);
+        Usuario usuario = usuarioDao.selectUsuarioById(id)
+                .orElseThrow(() -> new UsuarioNotFoundException(
+                        "Usuario con id [%s] no encontrado".formatted(id)
+                ));
         usuario.setNombre(updateRequest.getNombre());
         usuario.setApellido(updateRequest.getApellido());
-        this.usuarioDao.updateUsuario(id, updateRequest);
+        this.usuarioDao.updateUsuario(usuario);
     }
     public void deleteUsuario(Integer id) {
+        if (!usuarioDao.existsUsuarioById(id)) {
+            throw new UsuarioNotFoundException(
+                    "Usuario con id [%s] no encontrado".formatted(id)
+            );
+        }
         this.usuarioDao.deleteUsuario(id);
     }
     public List<Usuario> getUsuariosByStatus(UserStatus userStatus) {
